+++ src/tasks/shop-admin-tasks.ts
@@ -7,6 +7,7 @@ import { BulkEditProducts } from './shop-admin/Product/BulkEditProducts';
 import { BulkEditCustomers } from './shop-admin/Customers/BulkEditCustomers';
 import { AssignEntitiesToRule } from './shop-admin/Rule/AssignEntitiesToRule';
 import { CreateFlow } from './shop-admin/Flow/CreateFlow';
+import { LoginViaReviewsTab } from './shop-customer/Account/LoginViaReviewsTab';
 
 export const test = mergeTests(
     SaveProduct,
@@ -16,4 +17,5 @@ export const test = mergeTests(
     BulkEditCustomers,
     AssignEntitiesToRule,
     CreateFlow,
+    LoginViaReviewsTab,
 );